+++ codex-rs/protocol/src/protocol.rs
@@ -919,6 +919,7 @@ pub enum SessionSource {
     VSCode,
     Exec,
     Mcp,
+    SubAgent,
     #[serde(other)]
     Unknown,
 }
